updated all domain aggregares to seperate props and entity from aggregate file
