+++ packages/api-domain/src/domain/contexts/conversation/conversation/conversation.entity.ts
@@ -1,0 +1,18 @@
+import type { DomainSeedwork } from '@cellix/domain-seedwork';
+import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.entity.ts';
+import type { ItemListingEntityReference } from '../../listing/item/item-listing.entity.ts';
+
+export interface ConversationProps extends DomainSeedwork.DomainEntityProps {
+	sharer: Readonly<PersonalUserEntityReference>;
+	loadSharer: () => Promise<Readonly<PersonalUserEntityReference>>;
+	reserver: Readonly<PersonalUserEntityReference>;
+	loadReserver: () => Promise<Readonly<PersonalUserEntityReference>>;
+	listing: Readonly<ItemListingEntityReference>;
+	loadListing: () => Promise<Readonly<ItemListingEntityReference>>;
+	twilioConversationId: string;
+	createdAt: Date;
+	updatedAt: Date;
+}
+
+export interface ConversationEntityReference
+	extends Readonly<ConversationProps> {}
