+++ packages/api-domain/src/domain/contexts/conversation/conversation/conversation.uow.ts
@@ -1,7 +1,8 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
 import type { Passport } from '../../passport.ts';
 import type { ConversationRepository } from './conversation.repository.ts';
-import type { Conversation, ConversationProps } from './conversation.ts';
+import type { Conversation } from './conversation.ts';
+import type { ConversationProps } from './conversation.entity.ts';
 
 export interface ConversationUnitOfWork
 	extends DomainSeedwork.UnitOfWork<
