+++ packages/api-domain/src/domain/contexts/conversation/conversation/index.ts
@@ -1,8 +1,9 @@
 export { Conversation } from './conversation.ts';
+export type {} from './conversation.ts';
 export type {
 	ConversationProps,
 	ConversationEntityReference,
-} from './conversation.ts';
+} from './conversation.entity.ts';
 export type { ConversationRepository } from './conversation.repository.ts';
 export type { ConversationUnitOfWork } from './conversation.uow.ts';
 export { Message } from './message.entity.ts';
