+++ packages/api-domain/src/domain/contexts/listing/item/index.ts
@@ -1,8 +1,8 @@
-export {
-	ItemListing,
-	type ItemListingProps,
-	type ItemListingEntityReference,
-} from './item-listing.ts';
+export type {
+	ItemListingProps,
+	ItemListingEntityReference,
+} from './item-listing.entity.ts';
+export { ItemListing } from './item-listing.ts';
 export type { ItemListingRepository } from './item-listing.repository.ts';
 export type { ItemListingUnitOfWork } from './item-listing.uow.ts';
 export * as ItemListingValueObjects from './item-listing.value-objects.ts';
