+++ packages/api-domain/src/domain/contexts/listing/item/item-listing.repository.ts
@@ -1,6 +1,7 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
-import type { ItemListing, ItemListingProps } from './item-listing.ts';
-import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.ts';
+import type { ItemListing } from './item-listing.ts';
+import type { ItemListingProps } from './item-listing.entity.ts';
+import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.entity.ts';
 
 export interface ItemListingRepository<props extends ItemListingProps>
 	extends DomainSeedwork.Repository<ItemListing<props>> {
