+++ packages/api-domain/src/domain/contexts/listing/item/item-listing.uow.ts
@@ -1,7 +1,8 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
 import type { Passport } from '../../passport.ts';
-import type { ItemListing, ItemListingProps } from './item-listing.ts';
+import type { ItemListing } from './item-listing.ts';
 import type { ItemListingRepository } from './item-listing.repository.ts';
+import type { ItemListingProps } from './item-listing.entity.ts';
 
 export interface ItemListingUnitOfWork
 	extends DomainSeedwork.UnitOfWork<
