+++ packages/api-domain/src/domain/contexts/listing/listing.passport.ts
@@ -1,5 +1,5 @@
 import type { ListingVisa } from './listing.visa.ts';
-import type { ItemListingEntityReference } from './item/item-listing.ts';
+import type { ItemListingEntityReference } from './item/item-listing.entity.ts';
 
 export interface ListingPassport {
 	forItemListing(root: ItemListingEntityReference): ListingVisa;
