+++ packages/api-domain/src/domain/contexts/reservation-request/reservation-request/index.ts
@@ -1,7 +1,7 @@
-export {
-	ReservationRequest,
-	type ReservationRequestProps,
-	type ReservationRequestEntityReference,
-} from './reservation-request.ts';
+export { ReservationRequest } from './reservation-request.ts';
+export type {
+	ReservationRequestProps,
+	ReservationRequestEntityReference,
+} from './reservation-request.entity.ts';
 export type { ReservationRequestRepository } from './reservation-request.repository.ts';
 export type { ReservationRequestUnitOfWork } from './reservation-request.uow.ts';
