+++ packages/api-domain/src/domain/contexts/reservation-request/reservation-request/reservation-request.aggregate.test.ts
@@ -1,17 +1,15 @@
 import { describe, it, expect } from 'vitest';
-import {
-	ReservationRequest,
-	type ReservationRequestProps,
-} from './reservation-request.ts';
-import type { ItemListingEntityReference } from '../../listing/item/item-listing.ts';
-import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.ts';
+import { ReservationRequest } from './reservation-request.ts';
+import type { ReservationRequestProps } from './reservation-request.entity.ts';
+import type { ItemListingEntityReference } from '../../listing/item/item-listing.entity.ts';
+import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.entity.ts';
 // ...existing code...
 import {
 	ReservationRequestStates,
 	ReservationRequestStateValue,
 } from './reservation-request.value-objects.ts';
 import type { Passport } from '../../passport.ts';
-import type { PersonalUserRoleEntityReference } from '../../role/personal-user-role/personal-user-role.ts';
+import type { PersonalUserRoleEntityReference } from '../../role/personal-user-role/personal-user-role.entity.ts';
 import { PersonalUserRolePermissions } from '../../role/personal-user-role/personal-user-role-permissions.ts';
 // Minimal test-only mocks for missing domain value objects
 class PersonalUserAccountProfileLocation {
