+++ packages/api-domain/src/domain/contexts/reservation-request/reservation-request/reservation-request.entity.ts
@@ -1,0 +1,25 @@
+import type { DomainSeedwork } from '@cellix/domain-seedwork';
+import type { ItemListingEntityReference } from '../../listing/item/item-listing.entity.ts';
+import type { PersonalUserEntityReference } from '../../user/personal-user/personal-user.entity.ts';
+
+export interface ReservationRequestProps
+	extends DomainSeedwork.DomainEntityProps {
+	state: string;
+	reservationPeriodStart: Date;
+	reservationPeriodEnd: Date;
+	readonly createdAt: Date;
+	readonly updatedAt: Date;
+	readonly schemaVersion: string;
+	listing: Readonly<ItemListingEntityReference>;
+	loadListing(): Promise<ItemListingEntityReference>;
+	reserver: Readonly<PersonalUserEntityReference>;
+	loadReserver(): Promise<PersonalUserEntityReference>;
+	closeRequestedBySharer: boolean;
+	closeRequestedByReserver: boolean;
+}
+
+export interface ReservationRequestEntityReference
+	extends Readonly<Omit<ReservationRequestProps, 'listing' | 'reserver'>> {
+	readonly listing: Readonly<ItemListingEntityReference>;
+	readonly reserver: Readonly<PersonalUserEntityReference>;
+}
