+++ packages/api-domain/src/domain/contexts/reservation-request/reservation-request/reservation-request.uow.ts
@@ -1,10 +1,8 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
 import type { Passport } from '../../passport.ts';
-import type {
-	ReservationRequest,
-	ReservationRequestProps,
-} from './reservation-request.ts';
+import type { ReservationRequest } from './reservation-request.ts';
 import type { ReservationRequestRepository } from './reservation-request.repository.ts';
+import type { ReservationRequestProps } from './reservation-request.entity.ts';
 
 export interface ReservationRequestUnitOfWork
 	extends DomainSeedwork.UnitOfWork<
@@ -12,7 +10,7 @@ export interface ReservationRequestUnitOfWork
 			ReservationRequestProps,
 			ReservationRequest<ReservationRequestProps>,
 			ReservationRequestRepository<ReservationRequestProps>
-		>, 
+		>,
 		DomainSeedwork.InitializedUnitOfWork<
 			Passport,
 			ReservationRequestProps,
