+++ packages/api-domain/src/domain/contexts/role/personal-user-role/personal-user-role.repository.ts
@@ -1,8 +1,6 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
-import type {
-	PersonalUserRole,
-	PersonalUserRoleProps,
-} from './personal-user-role.ts';
+import type { PersonalUserRole } from './personal-user-role.ts';
+import type { PersonalUserRoleProps } from './personal-user-role.entity.ts';
 
 export interface PersonalUserRoleRepository<props extends PersonalUserRoleProps>
 	extends DomainSeedwork.Repository<PersonalUserRole<props>> {
