+++ packages/api-domain/src/domain/contexts/user/personal-user/index.ts
@@ -1,8 +1,8 @@
-export {
-	PersonalUser,
-	type PersonalUserEntityReference,
-	type PersonalUserProps,
-} from './personal-user.ts';
+export { PersonalUser } from './personal-user.ts';
+export type {
+	PersonalUserProps,
+	PersonalUserEntityReference,
+} from './personal-user.entity.ts';
 export type { PersonalUserRepository } from './personal-user.repository.ts';
 export type { PersonalUserUnitOfWork } from './personal-user.uow.ts';
 export type { PersonalUserAccountProps } from './personal-user-account.ts';
