+++ packages/api-domain/src/domain/contexts/user/personal-user/personal-user.uow.ts
@@ -1,7 +1,8 @@
 import type { DomainSeedwork } from '@cellix/domain-seedwork';
 import type { Passport } from '../../passport.ts';
-import type { PersonalUser, PersonalUserProps } from './personal-user.ts';
+import type { PersonalUser } from './personal-user.ts';
 import type { PersonalUserRepository } from './personal-user.repository.ts';
+import type { PersonalUserProps } from './personal-user.entity.ts';
 
 export interface PersonalUserUnitOfWork
 	extends DomainSeedwork.UnitOfWork<
