+++ packages/api-domain/src/domain/contexts/user/user.passport.ts
@@ -1,5 +1,5 @@
 import type { UserVisa } from './user.visa.ts';
-import type { PersonalUserEntityReference } from './personal-user/personal-user.ts';
+import type { PersonalUserEntityReference } from './personal-user/personal-user.entity.ts';
 
 export interface UserPassport {
 	forPersonalUser(root: PersonalUserEntityReference): UserVisa;
